tools: s3-tests runner script
Add an s3tests runner script.

This script wraps around the [s3-tests][1] and manages execution, log
and result collection.
The script executes setup and teardown for an s3gw instance for each
individual test in order to avoid cascading test failures in case of a
crash or corrupted database/file structure. It also collects both the
logs from s3gw and from nosetests and compiles a JSON report detailing
the individual outcomes of each test.
It will try to automatically determine when to use a runtime container,
a full s3gw container or just rely on the host to provide both the
binary and the runtime.

[1]: https://github.com/ceph/s3-tests.git

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>